* Coins List Page Complete * Detail Page Complete * Test scenarios add for repo's fetchCoins endpoint

+++ core/model/src/main/kotlin/com/ozantech/ozantechcaseapp/core/model/remote/response/CoinHistoryResponse.kt
@@ -15,6 +15,6 @@ data class CoinHistoryResponse(
         @Json(name = "price")
         val price: String? = null,
         @Json(name = "timestamp")
-        val timestamp: Long? = null
+        val timestamp: Float? = null
     )
 }
